+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 ## [Unreleased]
 
 ## RELEASE 1.0.7 - 2017-04-21
-### Fix
+### Fixed
 - Smart Fields - Smart fields are sent in the detail view request
 
 ## RELEASE 1.0.6 - 2017-04-14