+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## [Unreleased]
 ### Fixed
-- Smart Fields - Add smart field in belongsTo Fields.
+- Smart Fields - Serialize Smart Fields values for belongsTo association.
 
 ## RELEASE 1.1.0 - 2017-04-27
 ### Added