+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 # Change Log
 
 ## [Unreleased]
+### Added
+- Smart Fields - Add errors in console when user create smart field with syntax error
+
 ### Fixed
 - Smart Fields - Serialize Smart Fields values for belongsTo association.