+++ CHANGELOG.md
@@ -2,7 +2,8 @@
 
 ## [Unreleased]
 ### Added
-- Customization Errors - Add errors in console when user create Forest customization with syntax errors in code.
+- Customization Errors - Do not send the apimap when users create Forest customization with syntax errors in code.
+- Customization Errors - Add errors in the console when users create Forest customization with syntax errors in code.
 
 ### Fixed
 - Smart Fields - Serialize Smart Fields values for belongsTo association.